+++ .mocharc.js
@@ -4,11 +4,19 @@ process.env.NODE_ENV = 'test';
 
 global.expect = require('chai').expect;
 
+
+require('@babel/register')({
+  include: [
+    '../index.js',
+    '../test/**',
+    '../node_modules/bpmn-moddle-5/**'
+  ],
+});
+
 module.exports = {
   exit: true,
   recursive: true,
   reporter: 'spec',
-  require: ['@babel/register'],
   timeout: 1000,
 };